[Feat] 카테고리 API 추가
- 카테고리에 맞는 상품 조회, 상품에 카테고리 추가 API 구현

+++ pampam-service/src/main/java/com/example/pampam/utils/ProductType.java
@@ -8,23 +8,23 @@ import java.util.Map;
 @Getter
 @AllArgsConstructor
 public enum ProductType {
-    Vegetable(1,"채소"),
-    Nut(2, "견과류"),
-    Grain(3,"곡류"),
-    Mushroom(4, "버섯류"),
-    Fruit(5, "과일"),
-    SeaFood(6, "해산물"),
-    DriedFish(7, "건어물"),
-    Meat(8, "정육"),
-    MilkProducts(9, "유제품"),
-    Drink(10, "주류")
+    Vegetable(1L,"채소"),
+    Nut(2L, "견과류"),
+    Grain(3L,"곡류"),
+    Mushroom(4L, "버섯류"),
+    Fruit(5L, "과일"),
+    SeaFood(6L, "해산물"),
+    DriedFish(7L, "건어물"),
+    Meat(8L, "정육"),
+    MilkProducts(9L, "유제품"),
+    Drink(10L, "주류")
     ;
 
-    private final Integer idx;
+    private final Long idx;
     private final String type;
-    private static final Map<Integer, String> map = new HashMap<>();
+    private static final Map<Long, String> map = new HashMap<>();
 
-    public static Map<Integer, String> findType() {
+    public static Map<Long, String> findType() {
         for (ProductType value : ProductType.values()) {
             map.put(value.idx, value.type);
         }
